pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.119.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tabulate>=0.9.0",
    "uvicorn>=0.37.0",
]
